# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: EDOLE AFRICA ADMIN - Unified operational management platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /equipment:
    get:
      operationId: listEquipment
      tags: [equipment]
      summary: List all equipment
      parameters:
        - in: query
          name: categoryId
          schema:
            type: string
        - in: query
          name: status
          schema:
            type: string
        - in: query
          name: available
          schema:
            type: boolean
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Paginated equipment list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedEquipment"
    post:
      operationId: createEquipment
      tags: [equipment]
      summary: Create equipment item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEquipmentBody"
      responses:
        "201":
          description: Created equipment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Equipment"
  /equipment/{id}:
    get:
      operationId: getEquipment
      tags: [equipment]
      summary: Get equipment details
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Equipment details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Equipment"
    put:
      operationId: updateEquipment
      tags: [equipment]
      summary: Update equipment
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEquipmentBody"
      responses:
        "200":
          description: Updated equipment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Equipment"
    delete:
      operationId: deleteEquipment
      tags: [equipment]
      summary: Delete equipment
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /equipment/categories:
    get:
      operationId: listEquipmentCategories
      tags: [equipment]
      summary: List equipment categories
      responses:
        "200":
          description: List of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EquipmentCategory"
    post:
      operationId: createEquipmentCategory
      tags: [equipment]
      summary: Create an equipment category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEquipmentCategoryBody"
      responses:
        "201":
          description: Created category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EquipmentCategory"
...
  /conversations:
    get:
      operationId: listConversations
      tags: [messaging]
      summary: List conversations for current user
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Paginated conversations
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedConversations"
    post:
      operationId: createConversation
      tags: [messaging]
      summary: Create a conversation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateConversationBody"
      responses:
        "201":
          description: Created conversation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Conversation"
  /conversations/{id}:
    get:
      operationId: getConversation
      tags: [messaging]
      summary: Get conversation details
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Conversation details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Conversation"
  /conversations/{id}/messages:
    get:
      operationId: listMessages
      tags: [messaging]
      summary: List messages in a conversation
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: Paginated messages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedMessages"
    post:
      operationId: sendMessage
      tags: [messaging]
      summary: Send a message
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendMessageBody"
      responses:
        "201":
          description: Sent message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"

  # ── CALLS ─────────────────────────────────────────────────────────────────
  /calls:
    get:
      operationId: listCallSessions
      tags: [calls]
      summary: List call sessions
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Paginated call sessions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedCallSessions"
    post:
      operationId: createCallSession
      tags: [calls]
      summary: Start a call session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCallSessionBody"
      responses:
        "201":
          description: Created call session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CallSession"
  /calls/{id}:
    get:
      operationId: getCallSession
      tags: [calls]
      summary: Get call session details
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Call session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CallSession"
    put:
      operationId: updateCallSession
      tags: [calls]
      summary: Update call session (end, etc.)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCallSessionBody"
      responses:
        "200":
          description: Updated call session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CallSession"

  # ── DASHBOARD ─────────────────────────────────────────────────────────────
...
  /notifications:
    get:
      operationId: listNotifications
      tags: [notifications]
      summary: List notifications for current user
      parameters:
        - in: query
          name: unreadOnly
          schema:
            type: boolean
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Paginated notifications
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedNotifications"
  /notifications/{id}/read:
    put:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark notification as read
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Notification marked read
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
...
components:
  schemas:
...
    Equipment:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        code:
          type: string
        categoryId:
          type: string
        categoryName:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [available, rented, maintenance, retired]
        quantity:
          type: integer
        availableQuantity:
          type: integer
        dailyRate:
          type: number
        imageUrl:
          type: string
        location:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - status
        - quantity
        - createdAt
    CreateEquipmentBody:
      type: object
      properties:
        name:
          type: string
        code:
          type: string
        categoryId:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [available, rented, maintenance, retired]
        quantity:
          type: integer
        dailyRate:
          type: number
        imageUrl:
          type: string
        location:
          type: string
      required:
        - name
        - status
        - quantity
    EquipmentCategory:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        equipmentCount:
          type: integer
      required:
        - id
        - name
    CreateEquipmentCategoryBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
      required:
        - name
...
    PaginatedEquipment:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Equipment"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - data
        - total
        - page
        - limit

    # ── RENTAL ────────────────────────────────────────────────────────────────
...
    Conversation:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        type:
          type: string
          enum: [direct, group, project, client]
        projectId:
          type: string
        clientId:
          type: string
        lastMessage:
          type: string
        lastMessageAt:
          type: string
          format: date-time
        unreadCount:
          type: integer
        participants:
          type: array
          items:
            type: object
            properties:
              userId:
                type: string
              name:
                type: string
              avatarUrl:
                type: string
            required:
              - userId
              - name
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - type
        - unreadCount
        - createdAt
    CreateConversationBody:
      type: object
      properties:
        title:
          type: string
        type:
          type: string
          enum: [direct, group, project, client]
        participantIds:
          type: array
          items:
            type: string
        projectId:
          type: string
        clientId:
          type: string
      required:
        - type
        - participantIds
    Message:
      type: object
      properties:
        id:
          type: string
        conversationId:
          type: string
        senderId:
          type: string
        senderName:
          type: string
        senderAvatarUrl:
          type: string
        content:
          type: string
        attachmentUrl:
          type: string
        readBy:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - conversationId
        - senderId
        - content
        - createdAt
    SendMessageBody:
      type: object
      properties:
        content:
          type: string
        attachmentUrl:
          type: string
      required:
        - content
    PaginatedConversations:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Conversation"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - data
        - total
        - page
        - limit
    PaginatedMessages:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Message"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - data
        - total
        - page
        - limit

    # ── CALLS ─────────────────────────────────────────────────────────────────
    CallSession:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
          enum: [audio, video]
        status:
          type: string
          enum: [scheduled, active, ended]
        initiatorId:
          type: string
        initiatorName:
          type: string
        conversationId:
          type: string
        projectId:
          type: string
        roomUrl:
          type: string
        startedAt:
          type: string
          format: date-time
        endedAt:
          type: string
          format: date-time
        durationSeconds:
          type: integer
        participants:
          type: array
          items:
            type: object
            properties:
              userId:
                type: string
              name:
                type: string
              joinedAt:
                type: string
                format: date-time
              leftAt:
                type: string
                format: date-time
            required:
              - userId
              - name
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - type
        - status
        - createdAt
    CreateCallSessionBody:
      type: object
      properties:
        type:
          type: string
          enum: [audio, video]
        conversationId:
          type: string
        projectId:
          type: string
        participantIds:
          type: array
          items:
            type: string
      required:
        - type
    UpdateCallSessionBody:
      type: object
      properties:
        status:
          type: string
          enum: [scheduled, active, ended]
        endedAt:
          type: string
          format: date-time
    PaginatedCallSessions:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/CallSession"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - data
        - total
        - page
        - limit

    # ── DASHBOARD ─────────────────────────────────────────────────────────────
...
    Notification:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        title:
          type: string
        body:
          type: string
        type:
          type: string
        entityType:
          type: string
        entityId:
          type: string
        isRead:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - title
        - body
        - type
        - isRead
        - createdAt
    PaginatedNotifications:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Notification"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - data
        - total
        - page
        - limit